chore(quota): review-fix polish for #309
- rename rateLimitFullyStale → rateLimitHasStaleWindow (matches body
  semantics — any window past resets_at, not all)
- preserve sub-second precision in ForecastCap: multiply float by
  Second before converting to Duration instead of truncating first
- whitespace-tolerant boundary regex in unchainStatuslineCommand so a
  hand-edited wrap with extra spaces round-trips cleanly
- mtime-keyed package-level cache on the 3 auth-file reads
  InheritRateLimits does on every /api/v1/sessions hit
- clean up bind-and-discard idiom in readOpenCodeInheritKey
- extract ChipBucket / mergeIntoBuckets / resolveChipMode from the
  ~75-line quotaChipData getter for readability
- tests: wrap round-trip with single-quoted commands (5 subtests),
  unchain extra whitespace, computeMetrics stale-null E2E,
  readCodexAccountID error paths (5 subtests) + missing-file case

+++ core/domain/session/rate_limit.go
@@ -141,7 +141,12 @@ func ForecastCap(history []RateLimitSnapshot, now time.Time) *time.Time {
 		return &t
 	}
 	secondsToCap := remaining / ratePerSecond
-	eta := time.Unix(latest.SampledAt, 0).Add(time.Duration(secondsToCap) * time.Second)
+	// Convert the float seconds-until-cap to a Duration without first
+	// truncating the float: `Duration(secondsToCap) * Second` would
+	// cast 12.7 → 12 before multiplying, losing sub-second precision.
+	// Multiplying then casting preserves it (rounded by Duration's
+	// integer nanosecond representation).
+	eta := time.Unix(latest.SampledAt, 0).Add(time.Duration(secondsToCap * float64(time.Second)))
 	if imminent.ResetsAt > 0 && eta.After(time.Unix(imminent.ResetsAt, 0)) {
 		// Won't hit cap before the window rolls over.
 		return nil
